Added recurrence class..
